pyproject.toml
[project]
name = "nightshade-release"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip>=0.2.0",
    "diffusers>=0.36.0",
    "einops>=0.8.1",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
